Move fix problems to separate file
